refactor(Client): promote client factory to test fixture
The client factory was embedded in server main source. However, it was not used by it, and it import some heavyweight dependencies (graalvm js engine, jython).
As it serves for test purposes, we put it in project test fixtures, so we can use it in tests, and we will be able to use it in future dependant test modules.